+++ summoner-cli/src/Summoner/GhcVer.hs
@@ -39,6 +39,7 @@ data GhcVer
     | Ghc844
     | Ghc865
     | Ghc883
+    | Ghc8101
     deriving stock (Eq, Ord, Show, Enum, Bounded)
 
 -- | Converts 'GhcVer' into dot-separated string.
@@ -50,6 +51,7 @@ showGhcVer = \case
     Ghc844  -> "8.4.4"
     Ghc865  -> "8.6.5"
     Ghc883  -> "8.8.3"
+    Ghc8101 -> "8.10.1"
 
 {- | These are old GHC versions that are not working with default GHC versions
 when using Stack.
@@ -69,6 +71,7 @@ latestLts = \case
     Ghc844  -> "lts-12.26"
     Ghc865  -> "lts-14.27"
     Ghc883  -> "lts-15.5"
+    Ghc8101 -> "nightly-2020-06-29"
 
 -- | Represents PVP versioning (4 numbers).
 data Pvp = Pvp
@@ -91,6 +94,7 @@ baseVerPvp = \case
     Ghc844  -> Pvp 4 11 1 0
     Ghc865  -> Pvp 4 12 0 0
     Ghc883  -> Pvp 4 13 0 0
+    Ghc8101 -> Pvp 4 14 0 0
 
 -- | Returns corresponding @base@ version of the given GHC version.
 baseVer :: GhcVer -> Text